+++ src/components/AboutUsComp/Introduction.jsx
@@ -1,0 +1,22 @@
+import React from "react";
+
+const Introduction = () => {
+  return (
+    <div className="panel col-span-full md:col-start-1 md:col-end-3 h-auto">
+      <div className="w-full font-semibold text-xl p-3 bg-slate-100">Intro</div>
+      <div className="p-3 text-justify">
+        Dedicated, passionate, and accomplished Full Stack Developer with 9+
+        years of progressive experience working as an Independent Contractor for
+        Google and developing and growing my educational social network that
+        helps others learn programming, web design, game development,
+        networking. I’ve acquired a wide depth of knowledge and expertise in
+        using my technical skills in programming, computer science, software
+        development, and mobile app development to developing solutions to help
+        organizations increase productivity, and accelerate business
+        performance.
+      </div>
+    </div>
+  );
+};
+
+export default Introduction;
